Consolidate generator methos
